gtanotexlwifi: Sync fstabs with my LineageOS 20.0 gtaxl-common device tree
Since Android 13+, Android enables a feature, project quotas, on the
data filesystem if the 'quota' filesystem parameter is specified at all.
After that, data partition mounting (as read/write) without the quota
parameter specified, just as it wasn't here before this sync, fails.
This sync, of course, fixes that problem with additions of the quota
parameter.